+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0]
+- Add decorated formatter
+- Add decorator for level color changing
+- Add decorator for level emoji prefixes
+- Add codecov badge
+
 ## [0.1.0]
 - Add .travis.yml for CI support
 - Fix support for package manager